backend/requirements.txt
flask
flask-pymongo
flask-bcrypt
flask-cors
flask-socketio
PyJWT
python-dotenv
dnspython
cloudinary